Fix wrong tint color in FWTextComponentView
